Move the String category nodes from gcore/src/logic.rs to text/src/lib.rs
