+++ node-graph/nodes/gcore/src/lib.rs
@@ -3,7 +3,6 @@ pub mod context;
 pub mod context_modification;
 pub mod debug;
 pub mod extract_xy;
-pub mod logic;
 pub mod memo;
 pub mod ops;
 
@@ -13,6 +12,5 @@ pub use context::*;
 pub use context_modification::*;
 pub use debug::*;
 pub use extract_xy::*;
-pub use logic::*;
 pub use memo::*;
 pub use ops::*;
